requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.74.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
comtypes==1.4.13
distro==1.9.0
docstring_parser==0.17.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iso8601==2.1.0
jiter==0.12.0
multidict==6.7.0
numpy==2.2.6
opencv-python==4.12.0.88
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
pypiwin32==223
python-dotenv==1.2.1
pyttsx3==2.99
pyvts==0.3.3
pywin32==311
sniffio==1.3.1
starlette==0.50.0
twitchio==2.9.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
